Contracts and Operations no longer static
Static contracts and operations had issues with import comparison (data would load onto static objects even though the save wasn't actually being loaded). They are no longer static, but still typesafe with the same loading validation.
This also forced me to move gertActionObject and getActionFromTypeAndName back to the Bladeburner class.
