+++ src/Bladeburner/ui/TeamSizeButton.tsx
@@ -1,6 +1,5 @@
 import type { Bladeburner } from "../Bladeburner";
-import type { BlackOperation } from "../Actions/BlackOperation";
-import type { Operation } from "../Actions/Operation";
+import type { BlackOperation, Operation } from "../Actions";
 
 import React, { useState } from "react";
 import { TeamSizeModal } from "./TeamSizeModal";
