+++ src/Bladeburner/ui/TeamSizeModal.tsx
@@ -1,6 +1,5 @@
 import type { Bladeburner } from "../Bladeburner";
-import type { Operation } from "../Actions/Operation";
-import type { BlackOperation } from "../Actions/BlackOperation";
+import type { BlackOperation, Operation } from "../Actions";
 
 import React, { useState } from "react";
 import { dialogBoxCreate } from "../../ui/React/DialogBox";
